pyproject.toml
[project]
name = "django-rest-celery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-extensions>=4.1",
]
